requirements.txt
fastapi[all]
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
motor